% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertByAnnotationDB.R
\name{annotationDBMeta}
\alias{annotationDBMeta}
\title{Create meta file for AnnotationDB}
\usage{
annotationDBMeta(cacheDir)
}
\arguments{
\item{cacheDir}{cacheDir for phantasus}
}
\description{
\code{createES} function creates an rds file containing meta information of
provided sqlite files for AnnotationDB
}
\examples{
\dontrun{
annotationDBMeta('/var/phantasus/cache')
}

}
