% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPCA.R
\name{calcPCA}
\alias{calcPCA}
\title{Principal Component Analysis.}
\usage{
calcPCA(es, replacena = "mean")
}
\arguments{
\item{es}{an ExpressionSet object, should be normalized}

\item{replacena}{method for replacing NA values (mean by default)}
}
\value{
json with full description of the plot for plotly.js
}
\description{
\code{calcPCA} calculates PCA-matrix for the given ExpressionSet
    and returns this matrix encoded to JSON.
}
\examples{
\dontrun{
data(es)
calcPCA(es)
}
}
\keyword{internal}
