% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatedAnnotation.R
\name{calculatedAnnotation}
\alias{calculatedAnnotation}
\title{Create calculated annotation}
\usage{
calculatedAnnotation(es, operation)
}
\arguments{
\item{es}{ExpressionSet object.}

\item{operation}{Name of the operation to perform calculation}
}
\description{
\code{calculatedAnnotation} adds a column calculated by operation
}
