% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGEO.R
\name{checkGPLsFallback}
\alias{checkGPLsFallback}
\title{Check possible annotations for GEO Dataset.}
\usage{
checkGPLsFallback(name)
}
\arguments{
\item{name}{String, containing GEO identifier of the dataset.}
}
\value{
Vector of filenames serialized in JSON format.
    If there is only one GPL for that dataset, the function will
    return \code{name}.
}
\description{
\code{checkGPLs} returns GPL-names for
    the specified GEO identifier.
}
\examples{
\dontrun{
checkGPLs('GSE27112')
checkGPLs('GSE14308')
}
}
