% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGEO.R
\name{checkGSEType}
\alias{checkGSEType}
\title{Checks GSE to be supported}
\usage{
checkGSEType(name, destDir, combine = any)
}
\arguments{
\item{name}{GSE id, with optional GPL specification}

\item{destDir}{path to cache directory}

\item{combine}{function on how to combine results, when multiple platforms are present}
}
\value{
logical vector if the dataset is supported or not
}
\description{
Checks GSE to be supported
}
\keyword{internal}
