% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPreloaded.R
\name{checkPreloadedNames}
\alias{checkPreloadedNames}
\title{Check names inside preloaded file}
\usage{
checkPreloadedNames(name)
}
\arguments{
\item{name}{String, containing filename. Assuming
that in the directory with preloaded files \code{preloadedDir}
exists file \code{filename.rda} with list of ExpressionSets \code{ess}.}
}
\value{
Vector of names serialized in JSON format.
}
\description{
\code{checkPreloadedNames} checks names of ExpressionSets that
  are included in file \code{name}
}
