% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaPlot.R
\name{colMeansByGroups}
\alias{colMeansByGroups}
\title{Calculate column averages in row groups}
\usage{
colMeansByGroups(m, groups)
}
\arguments{
\item{m}{matrix n x m}

\item{groups}{vector of size n of numbers from 1 to k}
}
\value{
matrix k*m of column averages  by groups
}
\description{
Calculate column averages in row groups
}
\keyword{internal}
