% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createES.R
\name{createES}
\alias{createES}
\title{Create ExpressionSet}
\usage{
createES(data, pData, labelDescription, colNames, rowNames, symbol)
}
\arguments{
\item{data}{gene expression matrix}

\item{pData}{matrix with phenotypical data}

\item{labelDescription}{names of phenoData}

\item{colNames}{sample ids}

\item{rowNames}{gene ids}

\item{symbol}{gene symbol}
}
\description{
Function for creating ExpressionSet from parameters and exporting it to global environment
}
