% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGEO.R
\name{getArchs4Files}
\alias{getArchs4Files}
\title{Returns list of ARCHS4 hdf5 files with expression data}
\usage{
getArchs4Files(cacheDir)
}
\arguments{
\item{cacheDir}{base directory for cache}
}
\value{
list of .h5 files
}
\description{
Returns list of ARCHS4 hdf5 files with expression data
}
\keyword{internal}
