% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appConfig.R
\name{getPhantasusConf}
\alias{getPhantasusConf}
\title{Read Phantasus Config}
\usage{
getPhantasusConf(
  value = NULL,
  configName = Sys.getenv("R_CONFIG_ACTIVE"),
  file = file.path(tools::R_user_dir(package = "phantasus", which = "config"),
    "user.conf")
)
}
\arguments{
\item{value}{Value to retrieve from the config file.}

\item{configName}{R_CONFIG_ACTIVE value. If unset, "default".}

\item{file}{Location of the config file}
}
\description{
Read Phantasus Config
}
