% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaPlot.R
\name{gseaPlot}
\alias{gseaPlot}
\title{Returns path to an svg file with enrichment plot}
\usage{
gseaPlot(fData, fVarLabels, rankBy, selectedGenes, width, height)
}
\arguments{
\item{fData}{list of annotation columns}

\item{rankBy}{name of the numeric column used for gene ranking}

\item{selectedGenes}{indexes of selected genes (starting from one, in the order of fData)}

\item{width}{width of the image (in inches)}

\item{height}{height of the image (in inches)}

\item{fvarLabels}{vector of column names}
}
\value{
path to an svg file
}
\description{
Returns path to an svg file with enrichment plot
}
