% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isHSDS}
\alias{isHSDS}
\title{check if url  responding as HSDS server
TRUE - hsds
FALSE - web link but not working
NULL - not web link}
\usage{
isHSDS(url)
}
\arguments{
\item{url}{URL to check}
}
\description{
check if url  responding as HSDS server
TRUE - hsds
FALSE - web link but not working
NULL - not web link
}
\keyword{internal}
