% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{kmeans}
\alias{kmeans}
\title{KMeans}
\usage{
kmeans(es, columns = c(), rows = c(), k, replacena = "mean")
}
\arguments{
\item{es}{an ExpressionSet object}

\item{columns}{list of specified columns' indices (optional)}

\item{rows}{list of specified rows' indices (optional)}

\item{k}{expected number of clusters}

\item{replacena}{method for replacing NA values (mean by default)}
}
\value{
json, containing array of corresponding clusters
}
\description{
Function for performing k-means clusterisation on ExpressionSet.
}
