% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{limmaAnalysis}
\alias{limmaAnalysis}
\title{Differential Expression analysis}
\usage{
limmaAnalysis(es, rows = c(), columns = c(), fieldValues)
}
\arguments{
\item{es}{an ExpressionSet object, should be normalized}

\item{rows}{vector of specified rows' indices (optional)}

\item{columns}{vector of specified columns' indices (optional)}

\item{fieldValues}{vector of comparison values, matching categories' names with columns (must be equal length with columns' vector if specified)}
}
\value{
file with serialized information on differential expression matrix
}
\description{
Function for performing differential expression analysis. Returns serialized de-matrix.
}
