% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGEO.R
\name{loadFromARCHS4}
\alias{loadFromARCHS4}
\title{Loads expression data from ARCHS4 count files.
Only sapmles with counted expression are kept.
If es already containts expression data it is returned as is.}
\usage{
loadFromARCHS4(es, archs4_files)
}
\arguments{
\item{es}{ExpressionSet from GEO to check for expression in ARCHS4}

\item{archs4_files}{list of available .h5 files from ARCHS4 project}
}
\value{
either original es or an ExpressionSet with loaded count data from ARCHS4
}
\description{
Loads expression data from ARCHS4 count files.
Only sapmles with counted expression are kept.
If es already containts expression data it is returned as is.
}
