% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGEO.R
\name{loadGEO}
\alias{loadGEO}
\title{Load GEO Dataset}
\usage{
loadGEO(name, type = NA)
}
\arguments{
\item{name}{string, containing GEO identifier of the dataset}

\item{type}{"GSE" or "GDS"}
}
\value{
file with serialized information on ExpressionSet
}
\description{
Function for loading dataset from GEO, creating ExpressionSet from it and exporting it to global environment
}
