% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGEO.R
\name{parseCharacteristics}
\alias{parseCharacteristics}
\title{Parse characteristics columns somewhat intelligently}
\usage{
parseCharacteristics(pdata)
}
\arguments{
\item{pdata}{AnnotatedDataFrame with characteristics columns}
}
\value{
named data.frame with parsed annotations, rows are in the same order as the input
}
\description{
Parse characteristics columns somewhat intelligently
}
\keyword{internal}
