% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performKmeans.R
\name{performKmeans}
\alias{performKmeans}
\title{K-means clusterisation.}
\usage{
performKmeans(es, columns = c(), rows = c(), k, replacena = "mean")
}
\arguments{
\item{es}{ExpressionSet object.}

\item{columns}{List of specified columns' indices (optional).}

\item{rows}{List of specified rows' indices (optional).}

\item{k}{Expected number of clusters.}

\item{replacena}{Method for replacing NA values
in series matrix (mean by default)}
}
\value{
Vector of corresponding clusters, serialized to JSON.
}
\description{
\code{performKmeans} returns a vector of corresponding clusters for
    each gene from a given ExpressionSet.
}
\examples{
data(es)
performKmeans(es, k = 2)

}
