% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPreloaded.R
\name{preloadedDirExists}
\alias{preloadedDirExists}
\title{Check existence of phantasusPreloadedDir}
\usage{
preloadedDirExists()
}
\value{
Boolean value.
}
\description{
\code{preloadedDirExists} checks if there is specified
  directory with preloaded files.
}
