% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertByAnnotationDB.R
\name{queryAnnotationDBMeta}
\alias{queryAnnotationDBMeta}
\title{Get meta list for annotationDB files}
\usage{
queryAnnotationDBMeta()
}
\value{
meta info in JSON
}
\description{
\code{createES} Function reads an rds file containing meta information of provded
sqlite files for AnnotationDB
}
\examples{
\dontrun{
queryAnnotationDBMeta()
}

}
