% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.gct}
\alias{read.gct}
\title{Reads ExpressionSet from a GCT file.
Function is deprecated, please use phantasusLite:::readGct() instead}
\usage{
read.gct(...)
}
\arguments{
\item{...}{additional options for read.csv}

\item{gct}{Path to gct file}
}
\value{
ExpressionSet object
}
\description{
Only versions 1.2 and 1.3 are supported.
}
\examples{
read.gct(system.file("extdata", "centers.gct", package = "phantasus"))
}
