% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceInR.R
\name{reproduceInR}
\alias{reproduceInR}
\title{Reproduce session in R code}
\usage{
reproduceInR(sessionName, leaf = T, step = 0, savedEnv = new.env())
}
\arguments{
\item{sessionName}{String, OCPU session name}

\item{leaf}{Boolean, is it leaf (default = F)}

\item{step}{Integer, step of recursion (default = 0)}

\item{savedEnv}{Environment, where to store complex arguments (default = new.env())}
}
\value{
JSON with R code
}
\description{
Reproduce session in R code
}
\examples{
\dontrun{
  setwd(tempdir())
  reproduceInR('x039f1672026678');
}
}
