% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serveMorpheus}
\alias{serveMorpheus}
\title{Starts http server handling morpheus static files and opencpu}
\usage{
serveMorpheus(host, port, morpheusRoot)
}
\arguments{
\item{host}{host to listen}

\item{port}{port to listen}

\item{morpheusRoot}{path to static files with morpheus (on local file system)}
}
\description{
Starts http server handling morpheus static files and opencpu
}

