% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appConfig.R
\name{setupPhantasus}
\alias{setupPhantasus}
\title{Setup phantasus.
Read user config file ( or create default one) and fill \code{cache_root} using sources in \code{file}.}
\usage{
setupPhantasus(setup_name = "default", file = confFile("setup.yml"))
}
\arguments{
\item{setup_name}{name of config from \code{file}. If unset or not existed, "default".}

\item{file}{Location of the setup.yml file with setup parameters. If not existed use file from package}
}
\description{
Setup phantasus.
Read user config file ( or create default one) and fill \code{cache_root} using sources in \code{file}.
}
