% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGAM.R
\name{shinyGAMAnalysis}
\alias{shinyGAMAnalysis}
\title{Constructs data frame with gene annotations and submits it into Shiny GAM web-server}
\usage{
shinyGAMAnalysis(fData, fvarLabels)
}
\arguments{
\item{fData}{list of annotation columns}

\item{fvarLabels}{vector of column names}
}
\value{
URL for Shiny GAM
}
\description{
Constructs data frame with gene annotations and submits it into Shiny GAM web-server
}
