% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateARCHS4}
\alias{updateARCHS4}
\title{Update archs4 files.}
\usage{
updateARCHS4(
  cacheDir = file.path(getPhantasusConf("cache_folders")$rnaseq_counts, "archs4"),
  organism = c("all"),
  force = FALSE
)
}
\arguments{
\item{cacheDir}{file path to \bold{archs4} cache directory}

\item{organism}{vector which determines organisms to download: human, mouse, zoo or all as default.
Also can be a genus. Possible genus: \enumerate{ \item drosophila \item gallus \item bos \item caenorhabditis
\item danio \item rattus \item saccharomyces \item arabidopsis}}

\item{force}{logical value which let function replace current files}
}
\description{
Download archs4 or archs4zoo counts in \code{cacheDir}. If directory does not exists function makes nothing and produce corresponding warnings.
}
\keyword{internal}
