% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateARCHS4meta}
\alias{updateARCHS4meta}
\title{Update ARCHS4 meta files}
\usage{
updateARCHS4meta(
  archDir = file.path(getOption("phantasusCacheDir"), "counts/archs4")
)
}
\arguments{
\item{archDir}{path to directory with arch4 .h5 files.}
}
\description{
Creates \code{meta.txt} file, which describes typical archs4 and archs4Zoo files.
}
\details{
This function produces very specific "hardcoded" \code{meta.txt} file for arch4 and archs4ZOO counts colletctions.
See \code{\link{validateCountsCollection}} for more common information and \code{meta.txt}  file structure
}
\seealso{
\code{\link{validateCountsCollection}}
}
