% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateDEE2meta}
\alias{updateDEE2meta}
\title{Update DEE2 meta files}
\usage{
updateDEE2meta(
  destDir = file.path(getPhantasusConf("cache_folders")$rnaseq_counts, "dee2")
)
}
\arguments{
\item{destDir}{path to directory with DEE2 .h5 files.}
}
\description{
Creates \code{meta.txt} file, which describes typical dee2 files.
}
\details{
This function produces very specific "hardcoded" \code{meta.txt} file for dee2 counts colletction.
See \code{\link{validateCountsCollection}} for more common information and \code{meta.txt}  file structure
}
\seealso{
\code{\link{validateCountsCollection}}
}
