% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateCountsCollection}
\alias{validateCountsCollection}
\alias{\alias{validateCountsCollection}}
\alias{Function}
\alias{checks}
\alias{existing}
\alias{and}
\alias{structure}
\alias{of}
\alias{\code{meta.txt}}
\alias{file}
\alias{in}
\alias{specified}
\alias{counts}
\alias{folder.Also}
\alias{it}
\alias{accessibility}
\alias{datasets}
\alias{corresponding}
\alias{\code{.h5}}
\alias{files.}
\title{Check a counts collection}
\usage{
validateCountsCollection(collectionDir, verbose = FALSE)
}
\arguments{
\item{collectionDir}{path to directory with collection}

\item{verbose}{logical value which determines a content of  the output.}
}
\description{
Check a counts collection
}
\details{
\code{collectionDir} should contain a bunch of \code{.h5} files
 and a single \code{meta.txt}.  \code{meta.txt} is \code{.tsv}-like file where for each \code{.h5} exists a row wit columns:
\describe{ \item{file_name}{ name of \code{.h5} file in \code{collectionDir}.}
           \item{sample_id}{name of dataset in \code{file_name} which contains sample IDs (sample_geo_accession for example).}
           \item{gene_id}{name of dataset in \code{file_name} which contains ids for genes. For correct work this dataset should contain unique values.}
           \item{gene_id_type}{meaning of gene_id. This value determines name of gene annotation field in phantasus.}
           }
}
