% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertByAnnotationDB.R
\name{annotationDBMeta}
\alias{annotationDBMeta}
\title{Create meta file for AnnotationDB}
\usage{
annotationDBMeta(annotDir)
}
\arguments{
\item{annotDir}{path to folder with annotationDB sqlite files}
}
\value{
nothing
}
\description{
\code{annotationDBMeta} function creates txt files containing meta information of
provided sqlite files for AnnotationDB.
}
\examples{
\dontrun{
annotationDBMeta('/var/phantasus/cache')
}

}
