% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatedAnnotation.R
\name{calculatedAnnotation}
\alias{calculatedAnnotation}
\title{Create calculated annotation}
\usage{
calculatedAnnotation(
  es,
  operation,
  rows = c(),
  columns = c(),
  isColumns = FALSE,
  name = NULL
)
}
\arguments{
\item{es}{ExpressionSet object.}

\item{operation}{Name of the operation to perform calculation}

\item{rows}{List of specified rows' indices (optional), indices start from 0}

\item{columns}{List of specified columns' indices (optional),
indices start from 0#'}

\item{isColumns}{Apply fn to columns}

\item{name}{Name of the new annotation}
}
\value{
Nothing. Annotated dataset will be assigned to es in environment
}
\description{
\code{calculatedAnnotation} adds a column calculated by operation
}
\keyword{internal}
