% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertByAnnotationDB.R
\name{convertByAnnotationDB}
\alias{convertByAnnotationDB}
\title{Map indexes using Annotation DB}
\usage{
convertByAnnotationDB(
  es,
  dbName,
  columnName,
  columnType,
  keyType,
  otherOptions
)
}
\arguments{
\item{es}{source ExpressionSet}

\item{dbName}{name of AnnotationDB file}

\item{columnName}{name of column in featureData of source ExpressionSet}

\item{columnType}{Type of indexes in columnName}

\item{keyType}{Type of mapped indexes}

\item{otherOptions}{additional parameters for conversion. Currently only named boolean deleteDotVersion is not ignored.}
}
\value{
JSON object with a vector of converted IDs
}
\description{
\code{convertByAnnotationDB} function returns \code{keyType} ids from \code{dbName} mapped to \code{columnName} in \code{es}.
}
\examples{
\dontrun{
}

}
\keyword{internal}
