% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertByAnnotationDB.R
\name{convertByAnnotationDB}
\alias{convertByAnnotationDB}
\title{Map indexes using Annotation DB}
\usage{
convertByAnnotationDB(es, dbName, columnName, columnType, keyType)
}
\arguments{
\item{es}{source ExpressionSet}

\item{dbName}{name of AnnotationDB file}

\item{columnName}{name of column in featureData of source ExpressionSet}

\item{columnType}{Type of indexes in columnName}

\item{keyType}{Type of mapped indexes}
}
\description{
\code{createES} function creates an rds file containing meta information of
provided sqlite files for AnnotationDB
}
\examples{
\dontrun{
}


}
