% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appConfig.R
\name{createDockerConf}
\alias{createDockerConf}
\title{Creates default docker conf file
Function creates default docker user configuration file based on provided \code{setup_file}
or on default parameters if \code{setup_file} doesn't exist. If \code{user_conf_file} exists function does nothing.}
\usage{
createDockerConf(
  setup_file = confFile("setup.yaml"),
  user_conf_file = confFile("user.conf")
)
}
\arguments{
\item{setup_file}{name of config from \code{file}. If unset or not existed, "default".}

\item{user_conf_file}{Location of the setup.yml file with setup parameters. If not existed use file from package}
}
\description{
Creates default docker conf file
Function creates default docker user configuration file based on provided \code{setup_file}
or on default parameters if \code{setup_file} doesn't exist. If \code{user_conf_file} exists function does nothing.
}
