% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createES.R
\name{createES}
\alias{createES}
\title{Create ExpressionSet}
\usage{
createES(data, pData, varLabels, fData, fvarLabels)
}
\arguments{
\item{data}{gene expression matrix}

\item{pData}{matrix with phenotypical data}

\item{varLabels}{names of phenoData columns}

\item{fData}{matrix with feature data}

\item{fvarLabels}{names of featureData columns}
}
\description{
Function for creating ExpressionSet from parameters and exporting it to global environment
}
