% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es-data.R
\docType{data}
\name{es}
\alias{es}
\title{Example dataset}
\format{An object of class \code{ExpressionSet} with 20 rows and 5 columns.}
\usage{
data(es)
}
\description{
Small slice from GSE27112-GPL6103 for
    runnable examples.
}
\examples{
\dontrun{
data(es)
performKmeans(es, k = 2)
}

}
\keyword{datasets}
