% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaPlot.R
\name{gseaPlot}
\alias{gseaPlot}
\title{Returns path to an svg file with enrichment plot}
\usage{
gseaPlot(
  es,
  rankBy,
  selectedGenes,
  width,
  height,
  vertical = FALSE,
  addHeatmap = FALSE,
  showAnnotation = NULL,
  annotationColors = NULL,
  pallete = c("blue", "white", "red")
)
}
\arguments{
\item{es}{ExpressionSet object.}

\item{rankBy}{name of the numeric column used for gene ranking}

\item{selectedGenes}{indexes of selected genes (starting from one, in the order of fData)}

\item{width}{width of the image (in inches)}

\item{height}{height of the image (in inches)}

\item{vertical}{whether to use vertical orientation (default: FALSE)}

\item{addHeatmap}{whether to add an expression heatmap, sorted by rankBy (default: FALSE)}

\item{showAnnotation}{a name of column annotation to add to the heatmap, default: NULL (no annotation)}

\item{annotationColors}{a list of colors to use in annotation}

\item{pallete}{a vector of colors to draw heatmap}
}
\value{
path to an svg file
}
\description{
Returns path to an svg file with enrichment plot
}
