% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaPlot.R
\name{gseaPlot}
\alias{gseaPlot}
\title{Returns path to an svg file with enrichment plot}
\usage{
gseaPlot(es, rankBy, selectedGenes, width, height, vertical = FALSE)
}
\arguments{
\item{es}{ExpressionSet object.}

\item{rankBy}{name of the numeric column used for gene ranking}

\item{selectedGenes}{indexes of selected genes (starting from one, in the order of fData)}

\item{width}{width of the image (in inches)}

\item{height}{height of the image (in inches)}

\item{vertical}{whether to use vertical orientation (default: FALSE)}
}
\value{
path to an svg file
}
\description{
Returns path to an svg file with enrichment plot
}
