% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{limmaAnalysis}
\alias{limmaAnalysis}
\title{Differential Expression analysis.}
\usage{
limmaAnalysis(es, fieldValues)
}
\arguments{
\item{es}{ExpressionSet object. It should be normalized for
more accurate analysis.}

\item{fieldValues}{Vector of comparison values, mapping
categories' names to columns/samples}
}
\value{
Name of the file containing serialized de-matrix.
}
\description{
\code{limmaAnalysis} performs differential expression analysis
    from limma package and returns a ProtoBuf-serialized resulting
    de-matrix.
}
\examples{
\dontrun{
data(es)
limmaAnalysis(es, fieldValues = c("A", "A", "A", "B", "B"))
}
}
