% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{limmaAnalysis}
\alias{limmaAnalysis}
\title{Differential Expression analysis.}
\usage{
limmaAnalysis(es, rows = c(), columns = c(), fieldValues)
}
\arguments{
\item{es}{ExpressionSet object. It should be normalized for
more accurate analysis.}

\item{rows}{Vector of specified rows' indices (optional).}

\item{columns}{Vector of specified columns' indices (optional).}

\item{fieldValues}{Vector of comparison values, mapping
categories' names to columns/samples
(must be equal length with columns' vector if specified).}
}
\value{
Name of the file containing serialized de-matrix.
}
\description{
\code{limmaAnalysis} performs differential expression analysis
  from limma package and returns a ProtoBuf-serialized resulting
  de-matrix.
}
