% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPreloaded.R
\name{loadPreloaded}
\alias{loadPreloaded}
\title{Load GEO Dataset.}
\usage{
loadPreloaded(name)
}
\arguments{
\item{name}{String, containing filename. Assuming
that in the directory with preloaded files \code{preloadedDir}
exists file \code{filename.rda} with list of ExpressionSets \code{ess}.}
}
\value{
File with ProtoBuf-serialized ExpressionSet-s
    that were loaded from specified file.
}
\description{
\code{loadPreloaded} returns the file with serialized ExpressionSets using
    ProtoBuf, that were preloaded on server.
}
\keyword{internal}
