% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{Principal Component Analysis.}
\usage{
pcaPlot(es, columns = c(), rows = c(), replacena = "mean")
}
\arguments{
\item{es}{an ExpressionSet object, should be normalized}

\item{columns}{list of specified columns' indices (optional)}

\item{rows}{list of specified rows' indices (optional)}

\item{replacena}{method for replacing NA values (mean by default)}
}
\value{
json with full description of the plot for plotly.js
}
\description{
Function for creating json with full description
  of the pcaPlot for plotly.js.
}
