% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{Makes nice PCA plot for expression data}
\usage{
pcaPlot(es, columns = c(), rows = c(), c1, c2, size = "", colour = "",
  label = "", replacena = "mean")
}
\arguments{
\item{es}{an ExpressionSet object, should be normalized}

\item{c1}{a number of the first component to plot (numeric)}

\item{c2}{a number of the second component to plot (numeric)}
}
\description{
Makes nice PCA plot for expression data
}
\examples{
pcaPlot(es.norm, 1, 2) + aes(color=time)
}

