% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertByAnnotationDB.R
\name{queryAnnotationDBMeta}
\alias{queryAnnotationDBMeta}
\title{Get meta list for annotationDB files}
\usage{
queryAnnotationDBMeta()
}
\value{
meta info in JSON
}
\description{
\code{queryAnnotationDBMeta} Function reads txt meta files for provided
sqlite annotation databases.
}
\examples{
\dontrun{
queryAnnotationDBMeta()
}
}
\keyword{internal}
