% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.gct}
\alias{read.gct}
\title{Reads ExpressionSet from a GCT file.
Function is deprecated, please use phantasusLite:::readGct() instead}
\usage{
read.gct(...)
}
\arguments{
\item{...}{parameters for phantasusLite:::readGct() call}
}
\value{
ExpressionSet object
}
\description{
Reads ExpressionSet from a GCT file.
Function is deprecated, please use phantasusLite:::readGct() instead
}
