% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.gct}
\alias{read.gct}
\title{Reads ExpressionSet from GCT file.}
\usage{
read.gct(gct, ...)
}
\arguments{
\item{gct}{Path to gct file}

\item{...}{additional options for read.csv}
}
\value{
ExpressionSet object
}
\description{
Only versions 1.2 and 1.3 are supported.
}
