% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{servePhantasus}
\alias{servePhantasus}
\title{Serve phantasus.}
\usage{
servePhantasus(
  host = getPhantasusConf("host"),
  port = getPhantasusConf("port"),
  staticRoot = system.file("www/phantasus.js", package = "phantasus"),
  cacheDir = getPhantasusConf("cache_root"),
  preloadedDir = getPhantasusConf("preloaded_dir"),
  openInBrowser = getPhantasusConf("open_in_browser"),
  quiet = getPhantasusConf("quiet")
)
}
\arguments{
\item{host}{Host to listen.}

\item{port}{Port to listen.}

\item{staticRoot}{Path to static files with phantasus.js
(on local file system).}

\item{cacheDir}{Full path to cache directory.}

\item{preloadedDir}{Full path to directory with preloaded files.}

\item{openInBrowser}{Boolean value which states if application will be
automatically loaded in default browser.}

\item{quiet}{Boolean value which states whether the connection log should
be hidden (default: TRUE)}
}
\value{
Running instance of phantasus application.
}
\description{
\code{servePhantasus} starts http server handling phantasus static files
    and opencpu server.
}
\examples{
\dontrun{
servePhantasus()
}
}
