% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{servePhantasus}
\alias{servePhantasus}
\title{Serve phantasus.}
\usage{
servePhantasus(host = "0.0.0.0", port = 8000,
  staticRoot = system.file("www/phantasus.js", package = "phantasus"),
  cacheDir = tempdir(), preloadedDir = NULL, runInBackground = TRUE,
  openInBrowser = TRUE)
}
\arguments{
\item{host}{Host to listen.}

\item{port}{Port to listen.}

\item{staticRoot}{Path to static files with phantasus.js
(on local file system).}

\item{cacheDir}{Full path to cache directory.}

\item{preloadedDir}{Full path to directory with preloaded files.}

\item{runInBackground}{Boolean variable that states if
function must run in background.}

\item{openInBrowser}{Boolean variable that states if
application should be automatically opened in default browser.}
}
\value{
Running instance of phantasus application.
}
\description{
\code{servePhantasus} starts http server handling phantasus static files
    and opencpu server.
}
\examples{
\dontrun{
servePhantasus('0.0.0.0', 8000, cacheDir=file.path(getwd(), 'cache'))
}
}
