% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{servePhantasus}
\alias{servePhantasus}
\title{Starts http server handling morpheus static files and opencpu}
\usage{
servePhantasus(host, port, staticRoot = system.file("www/phantasus.js",
  package = "phantasus"), cacheDir = tempdir())
}
\arguments{
\item{host}{host to listen}

\item{port}{port to listen}

\item{staticRoot}{path to static files with phantasus.js (on local file system)}

\item{cacheDir}{full path to cache directory}
}
\description{
Starts http server handling morpheus static files and opencpu
}
\examples{
\dontrun{
servePhantasus('0.0.0.0', 8000, cacheDir=file.path(getwd(), 'cache'))
}
}
