% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGAM.R
\name{shinyGAMAnalysis}
\alias{shinyGAMAnalysis}
\title{Constructs data frame with gene annotations and submits it into Shiny GAM web-server}
\usage{
shinyGAMAnalysis(es)
}
\arguments{
\item{es}{Expression set object}
}
\value{
URL for Shiny GAM
}
\description{
Constructs data frame with gene annotations and submits it into Shiny GAM web-server
}
\keyword{internal}
