% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetES}
\alias{subsetES}
\title{Subsets es, if rows or columns are not specified, all are retained}
\usage{
subsetES(es, columns = c(), rows = c())
}
\arguments{
\item{es}{ExpressionSet object.#'}

\item{columns}{List of specified columns' indices (optional), indices start from 0#'}

\item{rows}{List of specified rows' indices (optional), indices start from 0}
}
\value{
`es`'s subset
}
\description{
Subsets es, if rows or columns are not specified, all are retained
}
