% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateCountsMeta}
\alias{updateCountsMeta}
\alias{\alias{updateCountsMeta}}
\alias{Creates}
\alias{\code{meta.rda}}
\alias{file}
\alias{which}
\alias{contain}
\alias{information}
\alias{about}
\alias{all}
\alias{samples}
\alias{in}
\alias{collections.}
\alias{Also}
\alias{function}
\alias{checks}
\alias{\code{priority.txt}}
\alias{file.}
\alias{This}
\alias{is}
\alias{used}
\alias{to}
\alias{manage}
\alias{collections}
\alias{with}
\alias{the}
\alias{same}
\alias{samples.}
\title{Update meta-data for counts collections}
\usage{
updateCountsMeta(
  counts_dir = file.path(getOption("phantasusCacheDir"), "counts"),
  force = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{counts_dir}{path to counts cache directory}

\item{force}{logical value wich lets function replace existing \code{meta.rda} file}

\item{verbose}{logical value which determines a content of  the output.}
}
\description{
Update meta-data for counts collections
}
\details{
First of all function checks validity of \code{priority.txt} file. \bold{Every} Collection should have \bold{unique} priority.
If \code{priority.txt} is not valid function creates new one, setting priorities for each subdirectory(=collection) equal to order in \code{list.dir} output.

Function updates \code{meta.rda} if this file is older than at least one \code{.h5} file in counts files.
\code{meta.rda} is \code{data.table} which is a result of union \code{data.table}s produced by \code{\link{getCountsMetaPart}} for each collection
}
\seealso{
{\code{\link{validateCountsCollection}},\code{\link{updateCountsMeta}}}
 @import data.table
}
