% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write.gct}
\alias{write.gct}
\title{Saves ExpressionSet to a GCT file (version 1.3).
Function is deprecated, please use phantasusLite:::writeGct() instead}
\usage{
write.gct(...)
}
\arguments{
\item{...}{parameters for phantasusLite:::writeGct() call}
}
\value{
Result of the closing file (as in `close()` function`)
}
\description{
Saves ExpressionSet to a GCT file (version 1.3).
Function is deprecated, please use phantasusLite:::writeGct() instead
}
